Added moment.js + Display how many days since start
Caved and added moment.js to avoid doing fuckery with UK date formats when it seems like Date wants a US one.
Displaying the days since first start in a temp place, it currently doesn't do much but will be used to calculate things in the future. Maybe it's worth database but I don't think using a local db for an application is a good idea.